packages/app_question_game/requirements.txt
#packaging
setuptools>=47.1.1,<48.0.0
wheel>=0.34.2,<0.35.0

# web-framework
flask>=1.1.2,<1.2.0

# db management
Flask-SQLAlchemy>=2.4.3,<2.5.0

#testing
pytest>=5.4.2,<6.0.0
